fix overflows while encoding large images
The data type that was used to store the memory required for input raw
buffers was int. This works for normal dimensions. For images where
width and height are very large, say greater than 32768, then rgb input
buffer size is 32768 * 32768 * 4. This computation exceeds integer max
causing issues. This change addresses this by assigning the type for
widths/heights/strides/ as unsigned int and buffer sizes as size_t

Test: ./ultrahdr_unit_test

+++ fuzzer/ultrahdr_dec_fuzzer.cpp
@@ -55,8 +55,8 @@ void UltraHdrDecFuzzer::process() {
   int bottom = mFdp.ConsumeIntegral<int16_t>();
 
   auto applyResize = mFdp.ConsumeBool();
-  int resizeWidth = mFdp.ConsumeIntegralInRange<int16_t>(-32, kMaxWidth + 128);
-  int resizeHeight = mFdp.ConsumeIntegralInRange<int16_t>(-32, kMaxHeight + 128);
+  int resizeWidth = mFdp.ConsumeIntegralInRange<int32_t>(-32, kMaxWidth + 128);
+  int resizeHeight = mFdp.ConsumeIntegralInRange<int32_t>(-32, kMaxHeight + 128);
 
   auto buffer = mFdp.ConsumeRemainingBytes<uint8_t>();
