+++ lib/include/ultrahdr/ultrahdr.h
@@ -127,9 +127,9 @@ struct jpegr_uncompressed_struct {
   // Pointer to the data location.
   void* data;
   // Width of the gain map or the luma plane of the image in pixels.
-  size_t width;
+  unsigned int width;
   // Height of the gain map or the luma plane of the image in pixels.
-  size_t height;
+  unsigned int height;
   // Color gamut.
   ultrahdr_color_gamut colorGamut;
 
@@ -140,7 +140,7 @@ struct jpegr_uncompressed_struct {
   // Stride of Y plane in number of pixels. 0 indicates the member is uninitialized. If
   // non-zero this value must be larger than or equal to luma width. If stride is
   // uninitialized then it is assumed to be equal to luma width.
-  size_t luma_stride = 0;
+  unsigned int luma_stride = 0;
   // Stride of UV plane in number of pixels.
   // 1. If this handle points to P010 image then this value must be larger than
   //    or equal to luma width.
@@ -148,7 +148,7 @@ struct jpegr_uncompressed_struct {
   //    or equal to (luma width / 2).
   // NOTE: if chroma_data is nullptr, chroma_stride is irrelevant. Just as the way,
   // chroma_data is derived from luma ptr, chroma stride is derived from luma stride.
-  size_t chroma_stride = 0;
+  unsigned int chroma_stride = 0;
   // Pixel format.
   uhdr_img_fmt_t pixelFormat = UHDR_IMG_FMT_UNSPECIFIED;
   // Color range.
@@ -162,9 +162,9 @@ struct jpegr_compressed_struct {
   // Pointer to the data location.
   void* data;
   // Used data length in bytes.
-  int length;
+  size_t length;
   // Maximum available data length in bytes.
-  int maxLength;
+  size_t maxLength;
   // Color gamut.
   ultrahdr_color_gamut colorGamut;
 };
